# improve download performance, minimal effort