--- a/lib/transfer.c
+++ b/lib/transfer.c
@@ -336,7 +336,7 @@ static CURLcode readwrite_data(struct Curl_easy *data,
     if((k->keepon & KEEP_RECV_PAUSE) || !(k->keepon & KEEP_RECV))
       break;
 
-  } while(maxloops-- && data_pending(data));
+  } while(maxloops--);
 
   if((maxloops <= 0) || data_pending(data)) {
     /* did not read until EAGAIN or there is still pending data, mark as